move plan-related logics into plan.go
