+++ x/farming/keeper/keeper.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	gogotypes "github.com/gogo/protobuf/types"
 	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -70,52 +69,6 @@ func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
 // SetParams sets the parameters for the farming module.
 func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
 	k.paramSpace.SetParamSet(ctx, &params)
-}
-
-// GetNextPlanIDWithUpdate returns and increments the global Plan ID counter.
-// If the global plan number is not set, it initializes it with value 1.
-func (k Keeper) GetNextPlanIDWithUpdate(ctx sdk.Context) uint64 {
-	var id uint64
-	store := ctx.KVStore(k.storeKey)
-
-	bz := store.Get(types.GlobalPlanIdKey)
-	if bz == nil {
-		// initialize the PlanId
-		id = 1
-	} else {
-		val := gogotypes.UInt64Value{}
-
-		err := k.cdc.Unmarshal(bz, &val)
-		if err != nil {
-			panic(err)
-		}
-
-		id = val.GetValue()
-	}
-	bz = k.cdc.MustMarshal(&gogotypes.UInt64Value{Value: id + 1})
-	store.Set(types.GlobalPlanIdKey, bz)
-	return id
-}
-
-func (k Keeper) decodePlan(bz []byte) types.PlanI {
-	acc, err := k.UnmarshalPlan(bz)
-	if err != nil {
-		panic(err)
-	}
-
-	return acc
-}
-
-// MarshalPlan protobuf serializes an Plan interface
-func (k Keeper) MarshalPlan(plan types.PlanI) ([]byte, error) { // nolint:interfacer
-	return k.cdc.MarshalInterface(plan)
-}
-
-// UnmarshalPlan returns an Plan interface from raw encoded plan
-// bytes of a Proto-based Plan type
-func (k Keeper) UnmarshalPlan(bz []byte) (types.PlanI, error) {
-	var acc types.PlanI
-	return acc, k.cdc.UnmarshalInterface(bz, &acc)
 }
 
 // GetCodec return codec.Codec object used by the keeper
